+++ extensions/myip/src/index.tsx
@@ -1,34 +1,42 @@
 import { Action, ActionPanel, Icon, List, useNavigation } from "@raycast/api";
-import axios from "axios";
-import { useEffect, useState } from "react";
+import { useFetch } from "@raycast/utils";
+import { address } from "ip";
+import { useState } from "react";
+
 import LookUp from "./lookup";
 import Torrent from "./torrent";
-import { address } from "ip";
-
-export type LoadingStatus = "loading" | "success" | "failure";
+import { headers } from "./util";
 
 export default function Command() {
-  const [status, setStatus] = useState<LoadingStatus>("loading");
-  const [ip, setIp] = useState("");
   const { pop } = useNavigation();
   const [localIp] = useState(() => address("public", "ipv4").toString());
 
-  useEffect(() => {
-    async function getIp() {
-      try {
-        const { data } = await axios.get("https://api64.ipify.org");
-        setIp(data);
-        setStatus("success");
-      } catch (error) {
-        setIp("Failure");
-        setStatus("failure");
-      }
-    }
-    getIp();
-  }, []);
+  const {
+    isLoading: isLoadingIPV6,
+    data: ipv6,
+    error: errorIPV6,
+    revalidate: revalidateIPV6,
+  } = useFetch<string>("https://api64.ipify.org", {
+    headers,
+    keepPreviousData: true,
+  });
+  const {
+    isLoading: isLoadingIPV4,
+    data: ipv4,
+    error: errorIPV4,
+    revalidate: revalidateIPV4,
+  } = useFetch<string>("https://api.ipify.org", {
+    headers,
+    keepPreviousData: true,
+  });
+
+  const revalidate = () => {
+    revalidateIPV6();
+    revalidateIPV4();
+  };
 
   return (
-    <List isLoading={status === "loading"}>
+    <List isLoading={isLoadingIPV6 || isLoadingIPV4}>
       <List.Item
         icon={Icon.Desktop}
         title={localIp}
@@ -41,6 +49,12 @@ export default function Command() {
                   pop();
                 }}
               />
+              <Action
+                title="Refresh"
+                onAction={() => revalidate()}
+                icon={Icon.Repeat}
+                shortcut={{ key: "r", modifiers: ["cmd"] }}
+              />
             </ActionPanel>
           )
         }
@@ -51,20 +65,26 @@ export default function Command() {
         ]}
       />
       <List.Item
-        subtitle={ip === "" ? "Loading..." : ""}
+        subtitle={!ipv6 && isLoadingIPV6 ? "Loading..." : errorIPV6 ? "Failed to load" : undefined}
         icon={Icon.Globe}
-        title={ip}
+        title={ipv6 || "Loading..."}
         actions={
-          status === "success" && (
+          !isLoadingIPV6 && !!ipv6 ? (
             <ActionPanel>
               <Action.CopyToClipboard
-                content={ip}
+                content={ipv6}
                 onCopy={() => {
                   pop();
                 }}
               />
+              <Action
+                title="Refresh"
+                onAction={() => revalidate()}
+                icon={Icon.Repeat}
+                shortcut={{ key: "r", modifiers: ["cmd"] }}
+              />
             </ActionPanel>
-          )
+          ) : null
         }
         accessories={[
           {
@@ -72,15 +92,39 @@ export default function Command() {
           },
         ]}
       />
-      {status === "success" && (
+      {!isLoadingIPV4 && !errorIPV4 && !!ipv4 ? (
+        <List.Item
+          subtitle={ipv4 === "" ? "Loading..." : errorIPV4 ? "Failed to load" : undefined}
+          icon={Icon.Globe}
+          title={ipv4 || "Loading..."}
+          actions={
+            !isLoadingIPV4 && !!ipv4 ? (
+              <ActionPanel>
+                <Action.CopyToClipboard
+                  content={ipv4}
+                  onCopy={() => {
+                    pop();
+                  }}
+                />
+              </ActionPanel>
+            ) : null
+          }
+          accessories={[
+            {
+              text: "Public IP address (IPV4)",
+            },
+          ]}
+        />
+      ) : null}
+      {!isLoadingIPV6 && !errorIPV6 && !!ipv6 ? (
         <>
           <List.Item
-            icon={ip === "" ? "" : Icon.Eye}
+            icon={ipv6 === "" ? "" : Icon.Eye}
             title=""
             subtitle="IP Lookup"
             actions={
               <ActionPanel>
-                <Action.Push title="IP Lookup" target={<LookUp ip={ip} />} icon={Icon.Eye} />
+                <Action.Push title="IP Lookup" target={<LookUp ip={ipv6} />} icon={Icon.Eye} />
               </ActionPanel>
             }
             accessories={[
@@ -90,12 +134,12 @@ export default function Command() {
             ]}
           />
           <List.Item
-            icon={ip === "" ? "" : Icon.Download}
+            icon={ipv6 === "" ? "" : Icon.Download}
             title=""
             subtitle="Torrent History"
             actions={
               <ActionPanel>
-                <Action.Push title="Torrent History" target={<Torrent ip={ip} />} icon={Icon.Download} />
+                <Action.Push title="Torrent History" target={<Torrent ip={ipv6} />} icon={Icon.Download} />
               </ActionPanel>
             }
             accessories={[
@@ -105,7 +149,7 @@ export default function Command() {
             ]}
           />
         </>
-      )}
+      ) : null}
     </List>
   );
 }